Added in seperation for output devices
