+++ arduino-main/src/util/mapper.h
@@ -12,7 +12,11 @@
 #include "../equipment/input/phsensor.h"
 #include "../equipment/input/sonar.h"
 #include "../equipment/input/temperature.h"
-#include "../equipment/output/output-devices.h"
+//#include "../equipment/output/output-devices.h"
+#include "../equipment/output/output.h"
+#include "../equipment/output/thruster.h"
+#include "../equipment/output/armgripper.h"
+#include "../equipment/output/armrotation.h"
 #include "constants.h"
